requirements.txt
botok
streamlit
st-annotated-text